requirements.txt
numpy == 1.19.5
opencv-python ==3.2.0.8
Pillow == 6.2.0
pandas == 1.1.5